# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DownloadControlBody": {
        "description": "暂停 / 取消请求体。",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "DownloadControlBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Meshforge API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/agent/models": {
      "get": {
        "description": "列出某个 provider 下可用的模型。\n\n* provider='openai'：GET {base_url}/models，带 Bearer key（尽力而为）。\n* provider='ollama'：查询 Ollama 的 /api/tags（为兼容保留）。\n\nReturns:\n    `{'models': [...]}`；任何失败都静默返回空列表（前端只用来填下拉框）。",
        "operationId": "list_models_agent_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "default": "ollama",
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "base_url",
            "required": false,
            "schema": {
              "default": "",
              "title": "Base Url",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "api_key",
            "required": false,
            "schema": {
              "default": "",
              "title": "Api Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ollama_url",
            "required": false,
            "schema": {
              "default": "http://localhost:11434",
              "title": "Ollama Url",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Models Agent Models Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Models",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/model/hf-download": {
      "get": {
        "description": "以 SSE 流式下载一个 HuggingFace Hub 模型。\n\n下载到 MODELS_DIR / model_id，并按扩展 manifest 里声明的过滤规则执行\n（hf_skip_prefixes / hf_include_prefixes）。\n\nSSE 数据事件：{\"percent\": 0-100, \"file\": \"...\", \"status\": \"...\"}\n终止事件：done / paused / cancelled / error\n\nArgs:\n    repo_id: HF 仓库 id（'owner/repo'）；必须与扩展 manifest 声明的\n        hfRepo 一致，未声明或不相符的仓库一律拒绝下载。\n    model_id: 扩展 id，决定落盘目录。\n    skip_prefixes: JSON 数组字符串，覆盖 manifest 里的排除前缀。\n    include_prefixes: JSON 数组字符串，覆盖 manifest 里的包含前缀。\n    x_hf_token: `X-HF-Token` 请求头（FastAPI 注入）——HF token 不再走\n        URL query，避免进入访问日志 / 浏览器历史 / 代理日志（文档 12.6 / 13.3）。",
        "operationId": "hf_download_model_hf_download_get",
        "parameters": [
          {
            "in": "query",
            "name": "repo_id",
            "required": true,
            "schema": {
              "title": "Repo Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "skip_prefixes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Skip Prefixes"
            }
          },
          {
            "in": "query",
            "name": "include_prefixes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include Prefixes"
            }
          },
          {
            "in": "header",
            "name": "X-HF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Hf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hf Download",
        "tags": [
          "model"
        ]
      }
    },
    "/model/hf-download/cancel": {
      "post": {
        "description": "取消指定模型的下载（流侧会顺带清掉 .part 文件）。",
        "operationId": "cancel_hf_download_model_hf_download_cancel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadControlBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Hf Download Model Hf Download Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Hf Download",
        "tags": [
          "model"
        ]
      }
    },
    "/model/hf-download/pause": {
      "post": {
        "description": "暂停指定模型的下载。\n\n只是置位 Event；实际的暂停发生在下载线程下一次检查控制信号时。\n因此 .part 文件与进度都会保留，可后续续传。",
        "operationId": "pause_hf_download_model_hf_download_pause_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadControlBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pause Hf Download Model Hf Download Pause Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Hf Download",
        "tags": [
          "model"
        ]
      }
    },
...
  }
}